backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
requests
python-multipart
openpyxl
pandas
pillow
pymupdf